1. Fixed a deadlock where next_transmission-Link::TimeStamp() was returning -1. This was because next_transmission was being incorrectly set due to truncation errors in token counting. The fix was to make the token count a long double instead of an int to do fine grained tracking.
2. wait_time_ns now returns a uint64_t to handle wait times of larger than 4 seconds , since 4*10^9 ns is the
max. limit of an int. This was causing weird truncation errors as well.

3. pkt_queue_occupancy is now public so we can inspect it and check if there is a deadlock ie :
A situation where we wait forever despite having packets in the queue.

To test this, I run a flood ping to fill the queue, then kill it, and then see if it still drains at the specified rate.
And it does, so that's good news for us.